refactor(user): reorganize user profile types and hooks
- move user types to dedicated directory with separate files
- create dedicated hooks directory for profile-related queries and mutations
- update imports across components to use new paths
- improve error logging in axios interceptor
- optimize profile data fetching with better query management
