+++ frontend/src/types/user/userProfileApiResponse.ts
@@ -1,0 +1,11 @@
+import type { User } from "./user";
+
+export interface UserProfileApiResponse {
+  message: string;
+  data: {
+    user: User;
+    supportingCount: number;
+    supportersCount: number;
+    isSupporting?: boolean;
+  };
+}
